Refactor cmd module
Split to run and run_dev modules
